Responsibilities were divided into active and deactive
